mv visualTags out as a component, and getStyles out as an interface
